Move collapse logic to SideNav
