+++ packages/side-nav/src/SideNav.tsx
@@ -1,16 +1,33 @@
-import { ReactNode, ComponentProps } from 'react';
+import { ReactNode, useRef } from 'react';
+import { useSpring, animated } from 'react-spring';
+import { Box, backgroundColorMap } from '@ag.ds-next/box';
+import {
+	LinkProps,
+	mapResponsiveProp,
+	mq,
+	tokens,
+	useElementSize,
+	usePrefersReducedMotion,
+	useToggleState,
+} from '@ag.ds-next/core';
 
 import { SideNavContainer } from './SideNavContainer';
 import { SideNavTitle } from './SideNavTitle';
 import { SideNavGroup } from './SideNavGroup';
 import { SideNavLink } from './SideNavLink';
-import { findBestMatch, useSideNavIds } from './utils';
-import { LinkProps } from '@ag.ds-next/core';
+import { SideNavCollapseButton } from './SideNavCollapseButton';
+import {
+	localPaletteVars,
+	variantMap,
+	SideNavVariant,
+	findBestMatch,
+	useSideNavIds,
+} from './utils';
 
 export type SideNavProps = LinkListProps & {
 	'aria-label'?: string;
 	collapseTitle: string;
-	variant?: ComponentProps<typeof SideNavContainer>['variant'];
+	variant?: SideNavVariant;
 	title: ReactNode;
 	titleLink: string; // TODO: should this be optional
 };
@@ -24,22 +41,71 @@ export function SideNav({
 	titleLink,
 	title,
 }: SideNavProps) {
-	const sideNavIds = useSideNavIds();
+	const { bodyId, buttonId, navId, titleId } = useSideNavIds();
+	const { hover } = variantMap[variant];
+	const [isOpen, onToggle] = useToggleState(false, true);
+	const ref = useRef<HTMLDivElement>(null);
+	const { height } = useElementSize(ref);
+
+	const prefersReducedMotion = usePrefersReducedMotion();
+	const animatedHeight = useSpring({
+		from: { height: 0 },
+		to: { height: isOpen ? height : 0 },
+		immediate: prefersReducedMotion,
+	});
 	return (
-		<SideNavContainer
-			aria-label={ariaLabel}
-			ids={sideNavIds}
-			variant={variant}
-			collapseTitle={collapseTitle}
-		>
-			<SideNavTitle
-				isCurrentPage={activePath === titleLink}
-				id={sideNavIds.titleId}
-				href={titleLink}
+		<SideNavContainer aria-label={ariaLabel} variant={variant}>
+			<SideNavCollapseButton
+				isOpen={isOpen}
+				onClick={onToggle}
+				ariaControls={bodyId}
+				variant={variant}
+				id={buttonId}
 			>
-				{title}
-			</SideNavTitle>
-			<LinkList activePath={activePath} items={items} />
+				{collapseTitle}
+			</SideNavCollapseButton>
+			<animated.div
+				id={bodyId}
+				aria-labelledby={buttonId}
+				style={animatedHeight}
+				css={{
+					overflow: 'hidden',
+					[tokens.mediaQuery.min.md]: {
+						// Overwrite the animated height
+						// for tablet/desktop sizes.
+						overflow: 'unset',
+						height: 'auto !important',
+					},
+				}}
+			>
+				<Box
+					ref={ref}
+					as="nav"
+					role="navigation"
+					aria-labelledby={titleId}
+					id={navId}
+					fontFamily="body"
+					paddingLeft={{ xs: 1, md: 0 }}
+					paddingRight={{ xs: 1, md: 0 }}
+					fontSize="sm"
+					lineHeight="default"
+					css={mq({
+						[localPaletteVars.hover]: mapResponsiveProp(
+							hover,
+							(t) => backgroundColorMap[t]
+						),
+					})}
+				>
+					<SideNavTitle
+						isCurrentPage={activePath === titleLink}
+						id={titleId}
+						href={titleLink}
+					>
+						{title}
+					</SideNavTitle>
+					<LinkList activePath={activePath} items={items} />
+				</Box>
+			</animated.div>
 		</SideNavContainer>
 	);
 }
