+++ packages/side-nav/src/SideNavContainer.tsx
@@ -1,45 +1,18 @@
-import { useRef } from 'react';
 import type { PropsWithChildren } from 'react';
-import { useSpring, animated } from 'react-spring';
-import { Box, backgroundColorMap } from '@ag.ds-next/box';
-import {
-	mapResponsiveProp,
-	mq,
-	tokens,
-	useElementSize,
-	usePrefersReducedMotion,
-	useToggleState,
-} from '@ag.ds-next/core';
-
-import { SideNavCollapseButton } from './SideNavCollapseButton';
-import { localPaletteVars, variantMap, SideNavVariant } from './utils';
+import { Box } from '@ag.ds-next/box';
+import { variantMap, SideNavVariant } from './utils';
 
 export type SideNavContainerProps = PropsWithChildren<{
-	collapseTitle?: string;
-	ids: { bodyId: string; buttonId: string; titleId: string; navId: string };
 	variant: SideNavVariant;
 	'aria-label': string;
 }>;
 
 export const SideNavContainer = ({
 	children,
-	collapseTitle,
-	ids: { bodyId, buttonId, titleId, navId },
 	variant,
 	'aria-label': ariaLabel,
 }: SideNavContainerProps) => {
-	const { palette, background, hover } = variantMap[variant];
-	const [isOpen, onToggle] = useToggleState(false, true);
-	const ref = useRef<HTMLDivElement>(null);
-	const { height } = useElementSize(ref);
-
-	const prefersReducedMotion = usePrefersReducedMotion();
-	const animatedHeight = useSpring({
-		from: { height: 0 },
-		to: { height: isOpen ? height : 0 },
-		immediate: prefersReducedMotion,
-	});
-
+	const { palette, background } = variantMap[variant];
 	return (
 		<Box
 			as="aside"
@@ -48,51 +21,7 @@ export const SideNavContainer = ({
 			background={background}
 			palette={palette}
 		>
-			<SideNavCollapseButton
-				isOpen={isOpen}
-				onClick={onToggle}
-				ariaControls={bodyId}
-				variant={variant}
-				id={buttonId}
-			>
-				{collapseTitle}
-			</SideNavCollapseButton>
-			<animated.div
-				id={bodyId}
-				aria-labelledby={buttonId}
-				role="region"
-				style={animatedHeight}
-				css={{
-					overflow: 'hidden',
-					[tokens.mediaQuery.min.md]: {
-						// Overwrite the animated height
-						// for tablet/desktop sizes.
-						overflow: 'unset',
-						height: 'auto !important',
-					},
-				}}
-			>
-				<Box
-					ref={ref}
-					as="nav"
-					role="navigation"
-					aria-labelledby={titleId}
-					id={navId}
-					fontFamily="body"
-					paddingLeft={{ xs: 1, md: 0 }}
-					paddingRight={{ xs: 1, md: 0 }}
-					fontSize="sm"
-					lineHeight="default"
-					css={mq({
-						[localPaletteVars.hover]: mapResponsiveProp(
-							hover,
-							(t) => backgroundColorMap[t]
-						),
-					})}
-				>
-					{children}
-				</Box>
-			</animated.div>
+			{children}
 		</Box>
 	);
 };
